@@ -4,31 +4,41 @@
     "bootstrap": "lerna bootstrap",
     "build": "yarn run clear && lerna run build",
     "clear": "lerna exec --parallel -- rm -rf lib",
-    "storybook": "start-storybook -p 9001",
-    "build-storybook": "build-storybook",
+    "type-check": "tsc --noEmit",
+    "storybook": "storybook dev -p 6006 --ci",
+    "build-storybook": "storybook build",
     "deploy-storybook": "storybook-to-ghpages"
   },
   "dependencies": {
+    "@emotion/react": "11.11.0",
+    "@emotion/styled": "11.11.0",
+    "@mui/material": "5.13.2",
     "array-move": "^2.2.1",
-    "linkify-it": "3.0.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "linkify-it": "4.0.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-dropzone": "^10.2.1",
     "react-image-lightbox": "5.1.4",
-    "react-sortable-hoc": "^1.11.0",
-    "tlds": "1.226.0",
+    "react-sortable-hoc": "^2.0.0",
+    "tlds": "1.238.0",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@babel/core": "7.16.0",
-    "@material-ui/core": "4.12.3",
-    "@storybook/react": "5.3.6",
-    "@types/react": "17.0.33",
-    "@types/react-dom": "17.0.10",
-    "@types/slate": "0.47.9",
+    "@babel/core": "7.21.8",
+    "@babel/preset-env": "7.21.5",
+    "@babel/preset-react": "7.18.6",
+    "@babel/preset-typescript": "7.21.5",
+    "@storybook/addon-essentials": "^7.0.14",
+    "@storybook/addon-interactions": "^7.0.14",
+    "@storybook/addon-links": "^7.0.14",
+    "@storybook/blocks": "^7.0.14",
+    "@storybook/react": "7.0.14",
+    "@storybook/react-webpack5": "^7.0.14",
+    "@storybook/testing-library": "^0.1.0",
+    "@types/react": "18.2.6",
+    "@types/react-dom": "18.2.4",
     "@typescript-eslint/eslint-plugin": "5.2.0",
     "@typescript-eslint/parser": "5.2.0",
-    "babel-loader": "8.2.3",
     "eslint": "8.1.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-import": "2.25.2",
@@ -36,14 +46,17 @@
     "eslint-plugin-prettier": "4.0.0",
     "eslint-plugin-react": "7.26.1",
     "eslint-plugin-react-hooks": "4.2.0",
+    "eslint-plugin-storybook": "0.6.12",
     "lerna": "4.0.0",
     "prettier": "2.4.1",
     "rimraf": "3.0.2",
     "rollup": "2.58.3",
     "rollup-plugin-sourcemaps": "0.6.3",
     "rollup-plugin-terser": "7.0.2",
-    "slate": "0.66.5",
-    "slate-react": "0.69.0",
+    "slate": "0.94.1",
+    "slate-react": "0.94.2",
+    "storybook": "^7.0.14",
+    "tsconfig-paths-webpack-plugin": "4.0.1",
     "typescript": "4.4.2"
   },
   "prettier": {
@@ -53,7 +66,9 @@
     "trailingComma": "all",
     "parser": "typescript"
   },
-  "workspaces": [
-    "packages/**"
-  ]
+  "workspaces": {
+    "packages": [
+      "packages/**"
+    ]
+  }
 }
